Moved find_stars to starfit package
