+++ src/starfit/find_stars.py
